requirements.txt
Flask==2.3.3
GitPython==3.1.37
numpy==1.24.3
scipy==1.11.2
sentence-transformers==2.2.2
transformers==4.32.1
torch==2.0.1
Werkzeug==2.3.7
